# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAKBFANTASY",
        "description": "Api SAKBFANTASY Documentation",
        "contact": {
            "email": "info@sakb.com"
        },
        "license": {
            "name": "Sakb",
            "url": "http://sakb-co.com.sa/"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/group_eldwry/{type_group}": {
            "get": {
                "tags": [
                    "group_eldwry"
                ],
                "summary": "get group eldwry of user",
                "operationId": "group_eldwry",
                "parameters": [
                    {
                        "name": "access-token",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type-dev",
                        "in": "header",
                        "description": "ios or android",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ios"
                        }
                    },
                    {
                        "name": "val-dev",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "sakb"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "header",
                        "description": "default lang is ar ( ar , en)",
                        "schema": {
                            "type": "string",
                            "default": "ar"
                        }
                    },
                    {
                        "name": "type_group",
                        "in": "path",
                        "description": "default type_group is classic ( classic , head)",
                        "schema": {
                            "type": "string",
                            "default": "classic"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
...
        "/api/v1/subeldwry": {
            "get": {
                "tags": [
                    "subeldwry_fixtures"
                ],
                "summary": "get all subeldwry",
                "operationId": "subeldwry",
                "parameters": [
                    {
                        "name": "type-dev",
                        "in": "header",
                        "description": "ios or android",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ios"
                        }
                    },
                    {
                        "name": "val-dev",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "sakb"
                        }
                    },
                    {
                        "name": "lang",
                        "in": "header",
                        "description": "default lang is ar ( ar , en)",
                        "schema": {
                            "type": "string",
                            "default": "ar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "error"
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}